Implement loading of IR caches

+++ engine/runtime/src/main/scala/org/enso/compiler/ModuleCache.scala
@@ -9,10 +9,11 @@ import org.bouncycastle.jcajce.provider.digest.SHA3
 import org.bouncycastle.util.encoders.Hex
 import org.enso.compiler.ModuleCache.ToMaskedPath
 import org.enso.compiler.core.IR
+import org.enso.interpreter.runtime.builtin.Builtins
 import org.enso.interpreter.runtime.{Context, Module}
 import org.enso.logger.masking.MaskedPath
 
-import java.io.{ByteArrayOutputStream, IOException, ObjectOutputStream}
+import java.io._
 import java.nio.charset.{Charset, StandardCharsets}
 import java.nio.file._
 import java.util.logging.Level
@@ -201,21 +202,32 @@ class ModuleCache(private val module: Module) {
   )(implicit logger: TruffleLogger): Option[ModuleCache.CachedModule] = {
     val metadataPath = getCacheMetadataPath(cacheRoot)
     val dataPath     = getCacheDataPath(cacheRoot)
+
     loadCacheMetadata(metadataPath) match {
       case Some(meta) =>
         val sourceDigestValid = computeSourceDigest() == meta.sourceHash
-        val blobDigestValid   = computeBlobDigest(dataPath) == meta.blobHash
+        val blobBytes         = dataPath.readAllBytes()
+        val blobDigestValid   = computeDigest(blobBytes) == meta.blobHash
         val compilationStage =
           Module.CompilationStage.valueOf(meta.compilationStage)
 
         if (sourceDigestValid && blobDigestValid) {
-          Some(
-            ModuleCache.CachedModule(
-              // TODO [AA] This is a dummy as we don't write the cache yet.
-              IR.Module(List(), List(), List(), None),
-              compilationStage
-            )
-          )
+          val bais: ByteArrayInputStream = new ByteArrayInputStream(blobBytes)
+          val ois: ObjectInputStream     = new ObjectInputStream(bais)
+
+          val readObject = ois.readObject()
+          ois.close()
+
+          readObject match {
+            case mod: IR.Module =>
+              Some(ModuleCache.CachedModule(mod, compilationStage))
+            case _ =>
+              logger.log(
+                ModuleCache.logLevel,
+                s"Module `${module.getName.toString}` was corrupt on disk."
+              )
+              None
+          }
         } else {
           logger.log(
             ModuleCache.logLevel,
@@ -273,19 +285,21 @@ class ModuleCache(private val module: Module) {
 
     def doDeleteAt(file: TruffleFile): Unit = {
       try {
-        if (file.isWritable) {
-          file.delete()
-          logger.log(
-            ModuleCache.logLevel,
-            s"Invalidated the cache at [${file.toMaskedPath.applyMasking()}]."
-          )
-        } else {
-          logger.log(
-            ModuleCache.logLevel,
-            s"Cannot invalidate the cache at " +
-            s"[${file.toMaskedPath.applyMasking()}]. " +
-            s"Cache location not writable."
-          )
+        if (file.exists()) {
+          if (file.isWritable) {
+            file.delete()
+            logger.log(
+              ModuleCache.logLevel,
+              s"Invalidated the cache at [${file.toMaskedPath.applyMasking()}]."
+            )
+          } else {
+            logger.log(
+              ModuleCache.logLevel,
+              s"Cannot invalidate the cache at " +
+              s"[${file.toMaskedPath.applyMasking()}]. " +
+              s"Cache location not writable."
+            )
+          }
         }
       } catch {
         case _: NoSuchFileException =>
@@ -311,23 +325,38 @@ class ModuleCache(private val module: Module) {
     *         module
     */
   private def getIrCacheRoots(context: Context): Option[ModuleCache.Roots] = {
-    context.getPackageOf(module.getSourceFile).toScala.map { pkg =>
-      val irCacheRoot    = pkg.getIrCacheRootForPackage(Info.ensoVersion)
-      val qualName       = module.getName
-      val localCacheRoot = irCacheRoot.resolve(qualName.path.mkString("/"))
+    if (module != context.getBuiltins.getModule) {
+      context.getPackageOf(module.getSourceFile).toScala.map { pkg =>
+        val irCacheRoot    = pkg.getIrCacheRootForPackage(Info.ensoVersion)
+        val qualName       = module.getName
+        val localCacheRoot = irCacheRoot.resolve(qualName.path.mkString("/"))
 
+        val distribution = context.getDistributionManager
+        val pathSegments = List(
+          pkg.namespace,
+          pkg.name,
+          pkg.config.version,
+          Info.ensoVersion
+        ) ++ qualName.path
+        val path = distribution.LocallyInstalledDirectories.irCacheDirectory
+          .resolve(pathSegments.mkString("/"))
+        val globalCacheRoot = context.getTruffleFile(path.toFile)
+
+        ModuleCache.Roots(localCacheRoot, globalCacheRoot)
+      }
+    } else {
       val distribution = context.getDistributionManager
       val pathSegments = List(
-        pkg.namespace,
-        pkg.name,
-        pkg.config.version,
+        Builtins.NAMESPACE,
+        Builtins.PACKAGE_NAME,
+        Info.ensoVersion,
         Info.ensoVersion
-      ) ++ qualName.path
+      ) ++ module.getName.path
       val path = distribution.LocallyInstalledDirectories.irCacheDirectory
         .resolve(pathSegments.mkString("/"))
       val globalCacheRoot = context.getTruffleFile(path.toFile)
 
-      ModuleCache.Roots(localCacheRoot, globalCacheRoot)
+      Some(ModuleCache.Roots(globalCacheRoot, globalCacheRoot))
     }
   }
 
@@ -359,18 +388,13 @@ class ModuleCache(private val module: Module) {
     * @return the hex string representing the digest
     */
   private def computeSourceDigest(): String = {
-    val sourceBytes = module.getSourceFile.readAllBytes()
-    computeDigest(sourceBytes)
-  }
+    val sourceBytes = if (module.getSource.hasBytes) {
+      module.getSource.getBytes.toByteArray
+    } else {
+      module.getSource.getCharacters.toString.getBytes(StandardCharsets.UTF_8)
+    }
 
-  /** Computes the SHA3-224 digest of the provided file.
-    *
-    * @param file the file to compute the digest of
-    * @return the hex string representing the digest of `file`
-    */
-  private def computeBlobDigest(file: TruffleFile): String = {
-    val blobBytes = file.readAllBytes()
-    computeDigest(blobBytes)
+    computeDigest(sourceBytes)
   }
 
   /** Computes the SHA3-224 digest of the provided byte array.
